+++ app_test_suite/steps/scenarios/simple.py
@@ -38,6 +38,7 @@ from app_test_suite.steps.test_types import (
 )
 
 TEST_APP_CATALOG_NAME: str = "chartmuseum"
+TEST_APP_CATALOG_NAMESPACE: str = "default"
 CONTEXT_KEY_APP_CR: str = "app_cr"
 CONTEXT_KEY_APP_CM_CR: str = "app_cm_cr"
 CHART_YAML = "Chart.yaml"
@@ -206,7 +207,12 @@ class SimpleTestScenario(BuildStep, ABC):
             raise ATSTestError("Can't establish connection to the new test cluster")
 
         # prepare app platform and upload artifacts
-        self._ensure_app_platform_ready(self._cluster_info.kube_config_path)
+        if not self._cluster_info.app_platform_ready:
+            logger.debug("App Platform not initialized, running `apptestctl`")
+            self._ensure_app_platform_ready(self._cluster_info.kube_config_path)
+            self._cluster_info.app_platform_ready = True
+        else:
+            logger.debug("App Platform already initialized, not running `apptestctl`")
         self._upload_chart_to_app_catalog(config, config.chart_file)
 
         try:
@@ -237,13 +243,24 @@ class SimpleTestScenario(BuildStep, ABC):
         )
 
         app_obj = self._deploy_chart(
-            app_name, app_version, deploy_namespace, app_config_file_path, TEST_APP_CATALOG_NAME
+            app_name,
+            app_version,
+            deploy_namespace,
+            app_config_file_path,
+            TEST_APP_CATALOG_NAME,
+            TEST_APP_CATALOG_NAMESPACE,
         )
         context[CONTEXT_KEY_APP_CR] = app_obj.app
         context[CONTEXT_KEY_APP_CM_CR] = app_obj.app_cm
 
     def _deploy_chart(
-        self, app_name: str, app_version: str, deploy_namespace: str, app_config_file_path: str, app_catalog_name: str
+        self,
+        app_name: str,
+        app_version: str,
+        deploy_namespace: str,
+        app_config_file_path: str,
+        app_catalog_name: str,
+        app_catalog_namespace: str,
     ) -> ConfiguredApp:
         config_values = None
         if app_config_file_path:
@@ -256,6 +273,7 @@ class SimpleTestScenario(BuildStep, ABC):
             app_name,
             app_version,
             app_catalog_name,
+            app_catalog_namespace,
             self._default_app_cr_namespace,
             deploy_namespace,
             config_values,